feat: get rid of scheduler and actions lambdas
