+++ .github/workflows/terraform.yml
@@ -44,7 +44,6 @@ jobs:
           SUBNET_B=App_Dev_azb_net
           SUBNET_DATA_A=Data_Dev_aza_net
           SUBNET_DATA_B=Data_Dev_azb_net
-          GITHUB_REPO=sso-terraform-dev
           KEYCLOAK_V2_DEV_URL=https://dev.sandbox.loginproxy.gov.bc.ca
           KEYCLOAK_V2_TEST_URL=https://test.sandbox.loginproxy.gov.bc.ca
           KEYCLOAK_V2_PROD_URL=https://sandbox.loginproxy.gov.bc.ca
@@ -82,7 +81,6 @@ jobs:
           SUBNET_B=App_Prod_azb_net
           SUBNET_DATA_A=Data_Prod_aza_net
           SUBNET_DATA_B=Data_Prod_azb_net
-          GITHUB_REPO=sso_terraform
           KEYCLOAK_V2_DEV_URL=https://dev.loginproxy.gov.bc.ca
           KEYCLOAK_V2_TEST_URL=https://test.loginproxy.gov.bc.ca
           KEYCLOAK_V2_PROD_URL=https://loginproxy.gov.bc.ca
@@ -198,7 +196,6 @@ jobs:
           subnet_b="${{ env.SUBNET_B }}"
           subnet_data_a="${{ env.SUBNET_DATA_A }}"
           subnet_data_b="${{ env.SUBNET_DATA_B }}"
-          gh_repo="${{ env.GITHUB_REPO }}"
           keycloak_v2_dev_url="${{ env.KEYCLOAK_V2_DEV_URL }}"
           keycloak_v2_test_url="${{ env.KEYCLOAK_V2_TEST_URL }}"
           keycloak_v2_prod_url="${{ env.KEYCLOAK_V2_PROD_URL }}"
