+++ app/page-partials/my-dashboard/TeamInfoTabs/index.tsx
@@ -41,9 +41,8 @@ import ModalContents from 'components/WarningModalContents';
 import InfoOverlay from 'components/InfoOverlay';
 import Grid from '@button-inc/bcgov-theme/Grid';
 import { Grid as SpinnerGrid } from 'react-loader-spinner';
-import SubmittedStatusIndicator from 'components/SubmittedStatusIndicator';
 import ServiceAccountsList from './ServiceAccountsList';
-import { InfoMessage } from '@app/components/MessageBox';
+import { ErrorMessage, InfoMessage } from '@app/components/MessageBox';
 import { Link } from '@button-inc/bcgov-theme';
 import { SurveyContext } from '@app/pages/_app';
 
@@ -644,17 +643,23 @@ function TeamInfoTabs({ alert, currentUser, team, loadTeams }: Props) {
                     </Grid.Col>
                     <Grid.Col span={6}>
                       {serviceAccountInProgress && (
-                        <Grid cols={10}>
-                          <Grid.Row gutter={[]}>
-                            <Grid.Col span={10} align={'center'}>
-                              {activeServiceAccount?.requester && (
-                                <Requester>Submitted by: {activeServiceAccount?.requester}</Requester>
-                              )}
-                              <SubTitle>CSS API Account will be provisioned in approx 20 min</SubTitle>
-                              <SubmittedStatusIndicator integration={serviceAccountInProgress} />
-                            </Grid.Col>
-                          </Grid.Row>
-                        </Grid>
+                        <div
+                          style={{
+                            display: 'inline-flex',
+                            margin: '20px 0 20px 0',
+                            background: '#FFCCCB',
+                            borderRadius: '5px',
+                          }}
+                        >
+                          <div style={{ padding: 5 }}>
+                            <ErrorMessage>
+                              Your request for an API account could not be completed. Please{' '}
+                              <Link external href="mailto:bcgov.sso@gov.bc.ca">
+                                contact the Pathfinder SSO Team
+                              </Link>
+                            </ErrorMessage>
+                          </div>
+                        </div>
                       )}
                     </Grid.Col>
                   </Grid.Row>
