+++ lambda/__tests__/15.sso-logs.test.ts
@@ -13,14 +13,13 @@ jest.mock('../app/src/grafana', () => {
   };
 });
 
-jest.mock('@lambda-app/controllers/requests', () => {
-  const original = jest.requireActual('@lambda-app/controllers/requests');
+jest.mock('../app/src/keycloak/integration', () => {
+  const original = jest.requireActual('../app/src/keycloak/integration');
   return {
     ...original,
-    processIntegrationRequest: jest.fn(() => true),
+    keycloakClient: jest.fn(() => Promise.resolve(true)),
   };
 });
-
 describe('Fetch SSO Logs', () => {
   const integration = getUpdateIntegrationData({ integration: { projectName: 'test_project' } });
