+++ lambda/__tests__/16.integration-metrics.test.ts
@@ -1,19 +1,8 @@
 import { TEAM_ADMIN_IDIR_USERID_01, TEAM_ADMIN_IDIR_EMAIL_01, postTeam } from './helpers/fixtures';
 import { cleanUpDatabaseTables, createMockAuth } from './helpers/utils';
-import { createTeam } from './helpers/modules/teams';
-import supertest from 'supertest';
-import app from './helpers/server';
-import { API_BASE_PATH } from './helpers/constants';
 import { buildIntegration } from './helpers/modules/common';
-import { findClientRole } from '@lambda-app/keycloak/users';
 import { Integration } from 'app/interfaces/Request';
-import {
-  createIntegration,
-  fetchMetrics,
-  getIntegration,
-  getIntegrations,
-  updateIntegration,
-} from './helpers/modules/integrations';
+import { fetchMetrics } from './helpers/modules/integrations';
 
 jest.mock('../app/src/grafana', () => {
   return {
@@ -23,31 +12,15 @@ jest.mock('../app/src/grafana', () => {
 
 jest.mock('@lambda-app/authenticate');
 
-jest.mock('@lambda-app/controllers/requests', () => {
-  const original = jest.requireActual('@lambda-app/controllers/requests');
+jest.mock('../app/src/keycloak/integration', () => {
+  const original = jest.requireActual('../app/src/keycloak/integration');
   return {
     ...original,
-    processIntegrationRequest: jest.fn(() => true),
-  };
-});
-
-jest.mock('../actions/src/github', () => {
-  return {
-    mergePR: jest.fn(),
+    keycloakClient: jest.fn(() => Promise.resolve(true)),
   };
 });
 
 jest.mock('@lambda-shared/utils/ches');
-
-jest.mock('@lambda-app/authenticate');
-
-jest.mock('@lambda-actions/authenticate', () => {
-  return {
-    authenticate: jest.fn(() => {
-      return Promise.resolve(true);
-    }),
-  };
-});
 
 describe('create/manage integration by authenticated user', () => {
   let integration: Integration;
@@ -60,8 +33,6 @@ describe('create/manage integration by authenticated user', () => {
         projectName: 'TEST CSS Metrics Tab',
         prodEnv: true,
         submitted: true,
-        planned: true,
-        applied: true,
       });
       integration = integrationRes.body;
     });
