+++ lambda/__tests__/jest.setup.js
@@ -2,6 +2,7 @@ const { createImportSpecifier } = require('typescript');
 const { sequelize, models } = require('../shared/sequelize/models/models');
 process.env.GH_SECRET = 'test';
 process.env.API_AUTH_SECRET = 'test';
+process.env.NODE_ENV = 'development';
 
 afterAll(async () => {
   return sequelize.close();
