--- lambda/actions/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "actions",
-  "version": "1.0.0",
-  "description": "Lambda function for receiving github actions requests",
-  "license": "Apache-2.0",
-  "author": "SSO Team",
-  "scripts": {
-    "compile": "../node_modules/.bin/tsc && ../node_modules/.bin/tsc-alias",
-    "build": "../node_modules/.bin/ncc build src/main.ts -o dist -m",
-    "copy-templates": "../node_modules/.bin/copyfiles -u 1 ../shared/templates/**/*.html ../shared/templates/*.html dist --flat"
-  },
-  "main": "index.js",
-  "devDependencies": {
-    "@octokit/types": "^6.23.0"
-  },
-  "dependencies": {
-    "cors": "^2.8.5",
-    "express": "^4.18.2",
-    "octokit": "^1.7.2",
-    "serverless-http": "^3.1.0"
-  },
-  "resolutions": {
-    "minimist": ">=1.2.6"
-  }
-}
