--- lambda/actions/src/authenticate.ts
@@ -1,9 +1,0 @@
-export const authenticate = async (headers) => {
-  const { Authorization, authorization } = headers || {};
-  const authHeader = Authorization || authorization;
-  if (!authHeader || authHeader !== process.env.GH_SECRET) {
-    return false;
-  }
-
-  return true;
-};
