--- lambda/actions/src/controllers/heartbeat.ts
@@ -1,6 +1,0 @@
-import { sequelize } from '@lambda-shared/sequelize/models/models';
-
-export const wakeUpAll = async () => {
-  const result = await sequelize.query('SELECT NOW()');
-  return result && result[0];
-};
