--- lambda/actions/src/interfaces.ts
@@ -1,10 +1,0 @@
-interface Headers {
-  [key: string]: string;
-}
-
-export interface Response {
-  statusCode?: number;
-  headers?: Headers;
-  isBase64Encoded?: boolean;
-  body?: string;
-}
