--- lambda/actions/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "./build",
-    "sourceMap": false,
-    "rootDir": "../../",
-    "baseUrl": "./",
-    "paths": {
-      "@app/*": ["../../app/*"],
-      "@lambda-app/*": ["../app/src/*"],
-      "@lambda-shared/*": ["../shared/*"]
-    }
-  },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "**/*.spec.ts", "**/*.test.ts"]
-}
