+++ lambda/jest.config.js
@@ -10,7 +10,6 @@ module.exports = {
   moduleNameMapper: {
     '^@app/(.*)$': '<rootDir>/../app/$1',
     '^@lambda-app/(.*)$': '<rootDir>/app/src/$1',
-    '^@lambda-actions/(.*)$': '<rootDir>/actions/src/$1',
     '^@lambda-shared/(.*)$': '<rootDir>/shared/$1',
     '^@lambda-css-api/(.*)$': '<rootDir>/css-api/src/$1',
   },
