+++ lambda/shared/templates/processing-time.html
@@ -1,5 +1,5 @@
 <p>
-  <strong>The expected processing time is 20 minutes</strong><br />
+  <strong>The expected processing time is 5 minutes</strong><br />
   Once the request is approved, you will receive another email that your JSON Client Installation is ready. If you are
   not the requester, this email serves only to notify you of the request status.
 </p>
