+++ lambda/tsconfig.json
@@ -9,7 +9,6 @@
       "*": ["../*"],
       "@app/*": ["../app/*"],
       "@lambda-app/*": ["./app/src/*"],
-      "@lambda-actions/*": ["./actions/src/*"],
       "@lambda-shared/*": ["./shared/*"],
       "@lambda-css-api/*": ["./css-api/src/*"]
     }
