+++ terraform/api-gateway.tf
@@ -11,36 +11,6 @@ resource "aws_api_gateway_rest_api" "sso_backend" {
   }
 }
 
-# Proxy put requests to /actions to the actions lambda
-resource "aws_api_gateway_resource" "actions" {
-  rest_api_id = aws_api_gateway_rest_api.sso_backend.id
-  parent_id   = aws_api_gateway_rest_api.sso_backend.root_resource_id
-  path_part   = "actions"
-}
-
-resource "aws_api_gateway_resource" "actions_proxy" {
-  rest_api_id = aws_api_gateway_rest_api.sso_backend.id
-  parent_id   = aws_api_gateway_resource.actions.id
-  path_part   = "{proxy+}"
-}
-
-resource "aws_api_gateway_method" "actions_proxy" {
-  rest_api_id   = aws_api_gateway_rest_api.sso_backend.id
-  resource_id   = aws_api_gateway_resource.actions_proxy.id
-  http_method   = "ANY"
-  authorization = "NONE"
-}
-
-resource "aws_api_gateway_integration" "actions" {
-  rest_api_id = aws_api_gateway_rest_api.sso_backend.id
-  resource_id = aws_api_gateway_method.actions_proxy.resource_id
-  http_method = aws_api_gateway_method.actions_proxy.http_method
-
-  integration_http_method = "POST"
-  type                    = "AWS_PROXY"
-  uri                     = aws_lambda_function.actions.invoke_arn
-}
-
 # Proxy requests to /api to the css-api lambda
 resource "aws_api_gateway_resource" "api" {
   rest_api_id = aws_api_gateway_rest_api.sso_backend.id
