--- terraform/cloudwatch.tf
@@ -1,18 +1,0 @@
-resource "aws_cloudwatch_event_rule" "scheduler" {
-  name                = "tf-apply-scheduler"
-  schedule_expression = "rate(5 minutes)"
-}
-
-resource "aws_cloudwatch_event_target" "scheduler" {
-  rule      = aws_cloudwatch_event_rule.scheduler.name
-  target_id = "scheduler"
-  arn       = aws_lambda_function.scheduler.arn
-}
-
-resource "aws_lambda_permission" "scheduler" {
-  statement_id  = "AllowExecutionFromCloudWatch"
-  action        = "lambda:InvokeFunction"
-  function_name = aws_lambda_function.scheduler.function_name
-  principal     = "events.amazonaws.com"
-  source_arn    = aws_cloudwatch_event_rule.scheduler.arn
-}
