+++ terraform/lambda-css-api.tf
@@ -41,10 +41,6 @@ resource "aws_lambda_function" "css_api" {
       KEYCLOAK_V2_PROD_CLIENT_SECRET = var.keycloak_v2_prod_client_secret
       REALM_REGISTRY_API             = var.realm_registry_api
       GH_ACCESS_TOKEN                = var.gh_access_token
-      GH_REPO                        = var.gh_repo
-      GH_WORKFLOW_ID                 = var.gh_workflow_id
-      GH_BRANCH                      = var.gh_branch
-      GH_OWNER                       = var.gh_owner
       CHES_API_ENDPOINT              = var.ches_api_endpoint
       CHES_TOKEN_ENDPOINT            = var.ches_token_endpoint
       CHES_PASSWORD                  = var.ches_password
